CCZ4: Rename NewConstraints -> Constraints
The "New" was just a GRChombo relic for backwards compatibility. We can
get rid of that now.
