+++ Tests/ConstraintsTest/ConstraintsTest.cpp
@@ -14,7 +14,7 @@
 #include "doctestCLIArgs.hpp"
 
 // Our includes
-#include "NewConstraints.hpp"
+#include "Constraints.hpp"
 
 // AMReX headers
 #include <AMReX.H>
